requirements.txt
# Image processing
scikit-image>=0.19.0
opencv-python>=4.5.1
Pillow>=8.0.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Utilities
imutils>=0.5.4
pandas>=1.3.0
requests>=2.26.0
natsort>=7.1.1
tqdm>=4.62.0

# Visualization
matplotlib>=3.4.0

# Optional: For model summary
torchinfo>=1.7.0
